GEODE-7019: Fix closing of idle connections in C++ Native client

+++ cppcache/src/ThinClientPoolDM.cpp
@@ -429,7 +429,7 @@ void ThinClientPoolDM::cleanStaleConnections(std::atomic<bool>& isRunning) {
       }
     }
 
-    size_t replaceCount =
+    int replaceCount =
         m_attrs->getMinConnections() - static_cast<int>(savelist.size());
 
     LOGDEBUG("Preserving %d connections", savelist.size());
